Adds Janelia_build feature
The Janelia source plugins were moved into separate directories, but retain their original installation names which conflict with other plugins. When doing a janelia build, we scan the plugin files with cmake scripts, and remove from installation those that will conflict with their Janelia counterparts. Since the Janelia pipeline is heavily dependent on the original naming, renaming the installation name of the Janelia plugins is not an option.